+++ packages/example-browserify/index.js
@@ -9,7 +9,7 @@ const app = express();
 const port = 30000;
 
 app.use(bodyParser.json());
-app.use(express.static(path.join(__dirname, 'www')));
+app.use(express.static(path.join(__dirname, 'public')));
 
 app.get('/', (req, res) => {
   res.sendFile('index.html');