+++ packages/example-node/images/index.js
@@ -1,6 +1,6 @@
 require('isomorphic-fetch');
 const qrcode = require('yaqrcode');
-const createReport = require('../../lib/indexNode');
+const createReport = require('docx-templates').default;
 
 createReport({
   template: process.argv[2],